models/UserModel.py
from extensions import db
from datetime import datetime
from dataclasses import dataclass


@dataclass
class UserModel(db.Model):
    id: int
    username: str

    __tablename__ = 'user'
    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
    username = db.Column(db.String(200), nullable=False, unique=True)
    email = db.Column(db.String(100), nullable=False, unique=True)
    password = db.Column(db.String(200), nullable=False)
    create_time = db.Column(db.DateTime, default=datetime.now)
